+++ tapes/1dc9c8c6b1c1a207d40cab92193e7b87-1.js
@@ -1,0 +1,37 @@
+var path = require("path");
+
+/**
+ * OPTIONS /ldp/db/tables
+ *
+ * connection: keep-alive
+ * host: folio-snapshot-load-okapi.dev.folio.org
+ * proxy-connection: keep-alive
+ * accept: * / *
+ * access-control-request-method: GET
+ * access-control-request-headers: content-type,x-okapi-tenant,x-okapi-token
+ * origin: http://localhost:3001
+ * user-agent: Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_6) AppleWebKit/537.36 (KHTML, like Gecko) Cypress/7.7.0 Chrome/89.0.4328.0 Electron/12.0.0-beta.14 Safari/537.36
+ * sec-fetch-mode: cors
+ * sec-fetch-site: same-site
+ * sec-fetch-dest: empty
+ * referer: http://localhost:3001/ldp
+ * accept-encoding: gzip
+ * accept-language: en-GB
+ */
+
+module.exports = function (req, res) {
+  res.statusCode = 204;
+
+  res.setHeader("date", "Thu, 30 Sep 2021 18:55:07 GMT");
+  res.setHeader("connection", "keep-alive");
+  res.setHeader("access-control-allow-origin", "*");
+  res.setHeader("access-control-allow-methods", "PUT,PATCH,DELETE,GET,POST");
+  res.setHeader("access-control-allow-headers", "content-type,X-Okapi-Tenant,X-Okapi-Token,Authorization,X-Okapi-Request-Id,X-Okapi-Module-Id");
+  res.setHeader("access-control-max-age", "7200");
+
+  res.setHeader("x-yakbak-tape", path.basename(__filename, ".js"));
+
+  res.end();
+
+  return __filename;
+};